requirements.txt
-e .
aws_cdk.aws_logs
aws_cdk.aws_iam
aws_cdk.aws_lambda
aws_cdk.aws_kinesis
aws_cdk.aws_kinesisfirehose
aws_cdk.aws_kinesisanalytics